+++ echoprobe.go
